SystemCode/frontend/requirements.txt
django~=3.2.8
requests~=2.26.0
python-dateutil~=2.8.2
django-extensions
pydotplus
pandas~=1.3.4
numpy~=1.21.2
scikit-learn~=1.0